[interpreter] Update Makefile to ignore output of `cd` (#1683)
The interpreter Makefile captures the output of cding into the unittest
directory and listing the unittest files it contains. Under some circumstances,
the `cd` command will print the new current directory, and that output was
incorrectly being interpreted as the path to a unittest, causing the build to
fail.

Fix this problem by redirecting the output of cd, if any, to /dev/null.

Fixes #1681.